+++ sdk/cloudhealth/azure-resourcemanager-cloudhealth/src/main/java/com/azure/resourcemanager/cloudhealth/models/EntityHistoryRequest.java
@@ -29,6 +29,17 @@ public final class EntityHistoryRequest implements JsonSerializable<EntityHistor
      */
     private OffsetDateTime endAt;
 
+    /*
+     * Maximum number of health state transitions to return per page. Defaults to 1000.
+     */
+    private Integer top;
+
+    /*
+     * An opaque string value that identifies the portion of the result set to be returned with the next operation. Must
+     * not be combined with startAt or endAt.
+     */
+    private String nextMarker;
+
     /**
      * Creates an instance of EntityHistoryRequest class.
      */
@@ -76,6 +87,48 @@ public final class EntityHistoryRequest implements JsonSerializable<EntityHistor
     }
 
     /**
+     * Get the top property: Maximum number of health state transitions to return per page. Defaults to 1000.
+     * 
+     * @return the top value.
+     */
+    public Integer top() {
+        return this.top;
+    }
+
+    /**
+     * Set the top property: Maximum number of health state transitions to return per page. Defaults to 1000.
+     * 
+     * @param top the top value to set.
+     * @return the EntityHistoryRequest object itself.
+     */
+    public EntityHistoryRequest withTop(Integer top) {
+        this.top = top;
+        return this;
+    }
+
+    /**
+     * Get the nextMarker property: An opaque string value that identifies the portion of the result set to be returned
+     * with the next operation. Must not be combined with startAt or endAt.
+     * 
+     * @return the nextMarker value.
+     */
+    public String nextMarker() {
+        return this.nextMarker;
+    }
+
+    /**
+     * Set the nextMarker property: An opaque string value that identifies the portion of the result set to be returned
+     * with the next operation. Must not be combined with startAt or endAt.
+     * 
+     * @param nextMarker the nextMarker value to set.
+     * @return the EntityHistoryRequest object itself.
+     */
+    public EntityHistoryRequest withNextMarker(String nextMarker) {
+        this.nextMarker = nextMarker;
+        return this;
+    }
+
+    /**
      * {@inheritDoc}
      */
     @Override
@@ -85,6 +138,8 @@ public final class EntityHistoryRequest implements JsonSerializable<EntityHistor
             this.startAt == null ? null : DateTimeFormatter.ISO_OFFSET_DATE_TIME.format(this.startAt));
         jsonWriter.writeStringField("endAt",
             this.endAt == null ? null : DateTimeFormatter.ISO_OFFSET_DATE_TIME.format(this.endAt));
+        jsonWriter.writeNumberField("top", this.top);
+        jsonWriter.writeStringField("nextMarker", this.nextMarker);
         return jsonWriter.writeEndObject();
     }
 
@@ -109,6 +164,10 @@ public final class EntityHistoryRequest implements JsonSerializable<EntityHistor
                 } else if ("endAt".equals(fieldName)) {
                     deserializedEntityHistoryRequest.endAt = reader
                         .getNullable(nonNullReader -> CoreUtils.parseBestOffsetDateTime(nonNullReader.getString()));
+                } else if ("top".equals(fieldName)) {
+                    deserializedEntityHistoryRequest.top = reader.getNullable(JsonReader::getInt);
+                } else if ("nextMarker".equals(fieldName)) {
+                    deserializedEntityHistoryRequest.nextMarker = reader.getString();
                 } else {
                     reader.skipChildren();
                 }