+++ sdk/netapp/azure-resourcemanager-netapp/src/main/java/com/azure/resourcemanager/netapp/implementation/VolumeImpl.java
@@ -23,8 +23,6 @@ import com.azure.resourcemanager.netapp.models.EncryptionKeySource;
 import com.azure.resourcemanager.netapp.models.FileAccessLogs;
 import com.azure.resourcemanager.netapp.models.GetGroupIdListForLdapUserRequest;
 import com.azure.resourcemanager.netapp.models.GetGroupIdListForLdapUserResponse;
-import com.azure.resourcemanager.netapp.models.LargeVolumeType;
-import com.azure.resourcemanager.netapp.models.LdapServerType;
 import com.azure.resourcemanager.netapp.models.ListQuotaReportResult;
 import com.azure.resourcemanager.netapp.models.ListReplicationsRequest;
 import com.azure.resourcemanager.netapp.models.MountTargetProperties;
@@ -42,7 +40,6 @@ import com.azure.resourcemanager.netapp.models.SmbAccessBasedEnumeration;
 import com.azure.resourcemanager.netapp.models.SmbNonBrowsable;
 import com.azure.resourcemanager.netapp.models.SvmPeerCommandResponse;
 import com.azure.resourcemanager.netapp.models.Volume;
-import com.azure.resourcemanager.netapp.models.VolumeLanguage;
 import com.azure.resourcemanager.netapp.models.VolumePatch;
 import com.azure.resourcemanager.netapp.models.VolumePatchPropertiesDataProtection;
 import com.azure.resourcemanager.netapp.models.VolumePatchPropertiesExportPolicy;
@@ -243,10 +240,6 @@ public final class VolumeImpl implements Volume, Volume.Definition, Volume.Updat
         return this.innerModel().ldapEnabled();
     }
 
-    public LdapServerType ldapServerType() {
-        return this.innerModel().ldapServerType();
-    }
-
     public Boolean coolAccess() {
         return this.innerModel().coolAccess();
     }
@@ -349,20 +342,12 @@ public final class VolumeImpl implements Volume, Volume.Definition, Volume.Updat
         return this.innerModel().isLargeVolume();
     }
 
-    public LargeVolumeType largeVolumeType() {
-        return this.innerModel().largeVolumeType();
-    }
-
     public String originatingResourceId() {
         return this.innerModel().originatingResourceId();
     }
 
     public Long inheritedSizeInBytes() {
         return this.innerModel().inheritedSizeInBytes();
-    }
-
-    public VolumeLanguage language() {
-        return this.innerModel().language();
     }
 
     public BreakthroughMode breakthroughMode() {
@@ -837,11 +822,6 @@ public final class VolumeImpl implements Volume, Volume.Definition, Volume.Updat
         return this;
     }
 
-    public VolumeImpl withLdapServerType(LdapServerType ldapServerType) {
-        this.innerModel().withLdapServerType(ldapServerType);
-        return this;
-    }
-
     public VolumeImpl withCoolAccess(Boolean coolAccess) {
         if (isInCreateMode()) {
             this.innerModel().withCoolAccess(coolAccess);
@@ -954,16 +934,6 @@ public final class VolumeImpl implements Volume, Volume.Definition, Volume.Updat
 
     public VolumeImpl withIsLargeVolume(Boolean isLargeVolume) {
         this.innerModel().withIsLargeVolume(isLargeVolume);
-        return this;
-    }
-
-    public VolumeImpl withLargeVolumeType(LargeVolumeType largeVolumeType) {
-        this.innerModel().withLargeVolumeType(largeVolumeType);
-        return this;
-    }
-
-    public VolumeImpl withLanguage(VolumeLanguage language) {
-        this.innerModel().withLanguage(language);
         return this;
     }