+++ sdk/translation/azure-ai-translation-text/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 2.0.1 (2026-07-01)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core-http-netty` from `1.16.4` to version `1.16.5`.
+- Upgraded `azure-core` from `1.58.0` to version `1.58.1`.
+
+
 ## 2.0.0 (2026-06-06)
 
 ### Features Added